update to properly handle outputs
