+++ phrenology/common/__init__.py
@@ -1,0 +1,1 @@
+#!/bin/env python3
